+++ lib/src/impl/keys_impl.dart
@@ -18,4 +18,15 @@ class KeyImpl implements KeyApi {
   String getPublicKey(String privateKey) {
     return Bip340Util.getPublicKey(privateKey);
   }
+
+  @override
+  bool isValidPrivateKey(String privateKey) {
+    try {
+      HexUtil.decode(privateKey);
+      return privateKey.length ==
+          64; // Check if the length of the decoded bytes is 32, and the length of the hex string is 64.
+    } catch (e) {
+      return false;
+    }
+  }
 }